Add - addTarea(){

+++ 14_API_REST/lista-tareas.js
@@ -1,8 +1,11 @@
 import { FetchService } from "./fetch-service.js";
+import {  MENSAJES } from "./mensajes.js"
 
 export class ListaTareas {
     constructor() {
         this.nodoListaTareas = document.querySelector('#lista')
+        this.nodoBtnAdd.addEventListener('click', this.addTarea.bind(this))
+        this.nodoNewTarea = document.querySelector('#inTarea')
         this.uRL = 'http://localhost:3000/tareas'
         this.aTareas = []
         this.fetchService = new FetchService()
@@ -13,15 +16,13 @@ export class ListaTareas {
         this.fetchService.send(this.uRL, {method: 'GET' })
             .then( data => {
                 this.aTareas = data
-                this.renderLista() 
+                this.renderLista()
             },
             error => {console.dir(error)}
             )
     }
     
     renderLista() {
-        console.log('RenderLista')
-        console.log('Tareas', this.aTareas)
         this.nodoListaTareas.innerHTML = ''
         let html = ''
         this.aTareas.forEach(
@@ -29,7 +30,7 @@ export class ListaTareas {
         )
         this.nodoListaTareas.innerHTML = html
         this.aNodosChecks = document.querySelectorAll('[name="is-completa"]')
-        this.aNodosBorrar = document.querySelectorAll('.borrar-tarea')
+        this.aNodosBorrar = document.querySelectorAll('.borrar-activo')
         this.aNodosChecks.forEach(
             item => item.addEventListener('change', this.checkTarea.bind(this))
         )
@@ -46,10 +47,18 @@ export class ListaTareas {
             <span class="nombre-tarea ${data.isComplete ? 'hecho' : '' }">
             ${data.name}</span>
             <span id="borrar-${data.id}" data-id="${data.id}"
-                class="borrar-tarea">🗑️</span>
+                class="borrar-tarea ${data.isComplete ? 'borrar-activo' : 'inactivo' }">🗑️</span>
             </li>
         `
         return htmlView
+    }
+
+    addTarea(){
+        if (!this.nodoNewTarea.value) {return}
+        let body = {
+            name: this.nodoNewTarea.value,
+            isComplete: false
+        }
     }
 
     checkTarea(oEv) {
@@ -58,29 +67,36 @@ export class ListaTareas {
         let datos = {
             isComplete : oEv.target.checked
         }
-        let url = this.url + '/' + oEv.target.dataset.id
-        this.fetchService.send(url, {method:'PATCH',
-            body: JSON.stringify(datos)
-        }).then(
-            response => {
-                console.log(response)
-            },
-            error => console.log(error)
-        )
-    }
-
-    borrarTarea(oEv) {
-        console.log(oEv.target.dataset.id)
-        // TODO Borar en Servicio Web
         let url = this.uRL + '/' + oEv.target.dataset.id
-        this.fetchService.send(url, {method: 'DELETE' })
-            .then(
-                data => {console.log(data)
+        let headers = new Headers()
+        headers.append("Content-Type", "application/json");
+        console.dir(headers)
+        this.fetchService.send(url, {
+                method: 'PATCH', 
+                headers : headers,
+                body: JSON.stringify(datos)
+            }).then( // () => this.getTareas()
+                    response => {
+                    console.log(response)
                     this.getTareas()
                 },
                 error => console.log(error)
             )
-
     }
 
+    borrarTarea(oEv) {
+        console.log(oEv.target.dataset.id)
+        if (!window.confirm( MENSAJES.listaTareas.confirmacion)) {return}
+        // TODO Borar en Servicio Web
+        let url = this.uRL + '/' + oEv.target.dataset.id
+        this.fetchService.send(url, {method: 'DELETE' })
+            .then(
+                data => { 
+                    console.log(data)
+                    this.getTareas() 
+                },
+                error => console.log(error)
+            )
+    }
+ 
 }
